Reorganized folder & separated the `CMakeLists.txt`.
