+++ CMakeLists.txt
@@ -5,15 +5,5 @@ project(Operations
 	VERSION 0.1
 	DESCRIPTION "Operations"
 	LANGUAGES CXX)
-add_executable (run main.cpp)
-
-set(DIR
-	${CMAKE_SOURCE_DIR}/operation/quantization
-	${CMAKE_SOURCE_DIR}/operation/conv
-	${CMAKE_SOURCE_DIR}/tensor)
-
-include_directories(${DIR})
-
-target_compile_options(run PUBLIC -Wall -Werror)
-target_include_directories(run PUBLIC ${DIR})
-
+	
+add_subdirectory(src)
