+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "quantization.hpp"
+#include "operation/quantization/quantization.hpp"
 #include "conv2d.hpp"
 #include <vector>
 #include <iostream>
@@ -26,8 +26,25 @@ void quantization_and_dequantization()
 	std::cout << dequantized_weights << "\n";
 }
 
+void conv2d()
+{
+	using TensorType = float;
+	Tensor<TensorType> inputs("inputs",
+							  {3, 3},
+							  {-1.0, -2.0, -3.0, -4.0, -5.0, -6.0, -7.0, -8.0, -9.0});
+	Tensor<TensorType> kernel_3d("kernel_3d", {3, 3}, "ones");
+	Tensor<TensorType> kernel_1d("kernel_1d", {1, 1}, "ones");
+
+	Tensor<TensorType> conv2d_3d_3d_output = conv::conv2d(inputs, kernel_3d, "conv2d_3d_3d_output");
+	std::cout << conv2d_3d_3d_output << "\n";
+
+	Tensor<TensorType> conv2d_3d_1d_output = conv::conv2d(inputs, kernel_1d, "conv2d_3d_1d_output");
+	std::cout << conv2d_3d_1d_output << "\n";
+}
+
 int main()
 {
-	quantization_and_dequantization();
+	// quantization_and_dequantization();
+	conv2d();
 	return 0;
 }
